Updating sample data and docs, moving shared template functions to snippets, minor changes to tests
