+++ src/index.js
@@ -12,7 +12,8 @@ const {
   BaseFHIRExtractor,
   CSVAdverseEventExtractor,
   CSVCancerDiseaseStatusExtractor,
-  CSVCancerRelatedMedicationExtractor,
+  CSVCancerRelatedMedicationAdministrationExtractor,
+  CSVCancerRelatedMedicationRequestExtractor,
   CSVClinicalTrialInformationExtractor,
   CSVConditionExtractor,
   CSVObservationExtractor,
@@ -85,7 +86,8 @@ module.exports = {
   BaseFHIRModule,
   CSVAdverseEventExtractor,
   CSVCancerDiseaseStatusExtractor,
-  CSVCancerRelatedMedicationExtractor,
+  CSVCancerRelatedMedicationAdministrationExtractor,
+  CSVCancerRelatedMedicationRequestExtractor,
   CSVClinicalTrialInformationExtractor,
   CSVConditionExtractor,
   CSVFileModule,
