+++ src/templates/snippets/subject.js
@@ -1,0 +1,11 @@
+const { reference } = require('./reference');
+
+function subjectTemplate({ id }) {
+  return {
+    subject: reference({ id, resourceType: 'Patient' }),
+  };
+}
+
+module.exports = {
+  subjectTemplate,
+};
